requirements.txt
requests
geoip2
python-dotenv
python-telegram-bot